endee-ai-assistant/backend/requirements.txt
# ─────────────────────────────────────────────
# AI Knowledge Assistant — Backend Dependencies
# ─────────────────────────────────────────────

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# File upload & form handling
python-multipart==0.0.9

# PDF parsing
PyMuPDF==1.24.3       # fitz — fast, lightweight PDF extraction

# Text chunking / NLP utility
langchain-text-splitters==0.2.2

# Embeddings (sentence-transformers runs locally, no API key needed)
sentence-transformers==2.7.0

# Endee vector database Python SDK
endee==0.1.18

# OpenAI (optional LLM backend — swap with HuggingFace if preferred)
openai==1.35.7

# Environment variable management
python-dotenv==1.0.1

# HTTP client used internally by the SDKs
httpx==0.28.1

# Data validation
pydantic==2.7.3
